WIP: Add fixtures for testing substitute_ids
