+++ anonymouus/tests/test_substitute_ids.py
@@ -1,6 +1,7 @@
 '''
 Test the core of the anonymoUUs package: substitution of strings
 '''
+
 import pytest
 from pathlib import Path
 import re
@@ -15,96 +16,85 @@ ids = [
     'casper',
     'caterpillar']
 
-# Options for providing keys and replacement values
-key_csv = Path.cwd()/'tests/test_data/keys.csv'
-key_dict = {
-'Jane Doe': 'aaaa',
-'Amsterdam':'bbbb',
-'j.doe@gmail.com':'cccc',
-re.compile('ca.*?er'):'dddd'
-} 
-
-# Anonymize instances 
-anym = Anonymize(keys)
-anym_case = Anonymize(keys,flags=re.IGNORECASE)
-anym_bounds = Anonymize(keys,use_word_boundaries=True)
-
 # Expected answers
 exp =        ['aaaa','JaneDoe','amsterdam','cccc','dddd','ddddpillar']
 exp_case =   ['aaaa','JaneDoe','bbbb','cccc','dddd','ddddpillar']
 exp_bounds = ['aaaa','JaneDoe','amsterdam','cccc','dddd','caterpillar']
 
-
-def test_substitute(anym,expected):
+@pytest.mark.parametrize('anym_object,expected',[
+                            ('reg',exp),
+                            ('case',exp_case),
+                            ('bounds',exp_bounds)
+                            ],
+                        indirect = True
+                        )
+def test_substitute(anym_object,expected):
     """Test substitution of strings"""
 
-    res = [anym._substitute_ids(s) for s in ids]
-    assert res == exp
+    res = [anym_object._substitute_ids(s) for s in ids]
+    assert res == expected
 
 
 
+# #@pytest.mark.parametrize("keys",[key_csv,key_dict])
+# @pytest.mark.parametrize("option,expected",[
+#                             (anym,exp),
+#                             (anym_case,exp_case),
+#                             (anym_bounds,exp_bounds)
+#                             ]
+#                         )
+# def test_substitute(option,expected):
+#     """Test substitution of strings"""
+
+#     anym = Anonymize(key_csv,option)
+#     res = [anym._substitute_ids(s) for s in ids]
+#     assert res == expected
 
 
-@pytest.fixture
-def anym_object(request):
-    """Create Anonymize object"""
-    return Anonymize(request.param)
+# def test_regular():
+#     """Test regular substitution of strings, without user options or flags"""
 
-@pytest.mark.parametrize(
-    'anym_object',
-    ([1, 2, 3], [2, 4, 6], [6, 8, 10]),
-    indirect=True
-)
+#     anym = Anonymize(key_csv)
 
+#     res = [anym._substitute_ids(s) for s in ids]
+#     exp = [
+#         'aaaa',
+#         'JaneDoe',
+#         'amsterdam',
+#         'cccc',
+#         'dddd',
+#         'ddddpillar']
 
+#     assert res == exp
 
+# def test_ignore_case(key_csv):
+#     """Test case-independent substitution of strings, without user options or flags"""
 
-@pytest.mark.parametrize('keys', [key_csv,key_dict])
-def test_regular(keys):
-    """Test regular substitution of strings, without user options or flags"""
+#     anym = Anonymize(key_csv,flags=re.IGNORECASE)
 
-    anym = Anonymize(keys)
+#     res = [anym._substitute_ids(s) for s in ids]
+#     exp = [
+#         'aaaa',
+#         'JaneDoe',
+#         'bbbb',
+#         'cccc',
+#         'dddd',
+#         'ddddpillar']
 
-    res = [anym._substitute_ids(s) for s in ids]
-    exp = [
-        'aaaa',
-        'JaneDoe',
-        'amsterdam',
-        'cccc',
-        'dddd',
-        'ddddpillar']
+#     assert res == exp
 
-    assert res == exp
+# def test_word_boundaries(key_csv):
+    # """Test case-independent substitution of strings, without user options or flags"""
 
-@pytest.mark.parametrize('keys', [key_csv,key_dict])
-def test_ignore_case(keys):
-    """Test case-independent substitution of strings, without user options or flags"""
+    # anym = Anonymize(key_csv,use_word_boundaries=True)
 
-    anym = Anonymize(keys,flags=re.IGNORECASE)
+    # res = [anym._substitute_ids(s) for s in ids]
+    # exp = [
+    #     'aaaa',
+    #     'JaneDoe',
+    #     'amsterdam',
+    #     'cccc',
+    #     'dddd',
+    #     'caterpillar']
 
-    res = [anym._substitute_ids(s) for s in ids]
-    exp = [
-        'aaaa',
-        'JaneDoe',
-        'bbbb',
-        'cccc',
-        'dddd',
-        'ddddpillar']
-
-    assert res == exp
-
-def test_word_boundaries(keys):
-    """Test case-independent substitution of strings, without user options or flags"""
-
-    anym = Anonymize(keys,use_word_boundaries=True)
-
-    res = [anym._substitute_ids(s) for s in ids]
-    exp = [
-        'aaaa',
-        'JaneDoe',
-        'amsterdam',
-        'cccc',
-        'dddd',
-        'caterpillar']
-
-    assert res == exp
+    # assert res == exp
